コーディング規則変更

+++ others/modcal.hpp
@@ -3,10 +3,10 @@
 
 /// @brief x^n (mod m) を返す
 template<typename T=ll>
-T ModPow(T x, T n, T mod){
+T ModPow(T x, T n, T mod) {
     ll ret=1;
-    if(typeid(T)==typeid(ll)&&mod>INF*2) return ModPow<__int128_t>(x,n,mod);
-    while(n>0){
+    if(typeid(T)==typeid(ll)&&mod>INF*2) return ModPow<LL>(x,n,mod);
+    while(n>0) {
         if(n&1) (ret*=x)%=mod;
         (x*=x)%=mod;
         n>>=1;
@@ -15,9 +15,9 @@ T ModPow(T x, T n, T mod){
 }
 
 /// @brief x^(-1) (mod m) を返す
-ll ModInv(ll a, ll m){
+ll ModInv(ll a, ll m) {
     ll b=m,u=1,v=0;
-    while(b){
+    while(b) {
         ll t=a/b;
         a-=t*b; swap(a,b);
         u-=t*v; swap(u,v);
